# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: OpenAgentMail API
  description: |
    Email API for AI agents. Create inboxes, send and receive email, and react to events in real-time.
    
    ## Authentication
    All endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_api_key>
    ```
    
    ## Pagination
    List endpoints use cursor-based pagination with `limit` and `page_token` parameters.
    
    ## Idempotency
    Use `client_id` to make POST requests idempotent.
  version: 0.1.0
  contact:
    name: OpenAgentMail
    url: https://github.com/openagentmail
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.openagentmail.com/v0
    description: Production
  - url: http://localhost:3000/v0
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /pods:
    get:
      tags: [Pods]
      summary: List pods
      operationId: listPods
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: List of pods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Pods]
      summary: Create pod
      operationId: createPod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePodRequest"
      responses:
        "201":
          description: Pod created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /pods/{pod_id}:
    get:
      tags: [Pods]
      summary: Get pod
      operationId: getPod
      parameters:
        - $ref: "#/components/parameters/podId"
      responses:
        "200":
          description: Pod details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pod"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /inboxes:
    get:
      tags: [Inboxes]
      summary: List inboxes
      operationId: listInboxes
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
      responses:
        "200":
          description: List of inboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Inboxes]
      summary: Create inbox
      operationId: createInbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInboxRequest"
      responses:
        "201":
          description: Inbox created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /inboxes/{inbox_id}:
    get:
      tags: [Inboxes]
      summary: Get inbox
      operationId: getInbox
      parameters:
        - $ref: "#/components/parameters/inboxId"
      responses:
        "200":
          description: Inbox details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Inboxes]
      summary: Delete inbox
      operationId: deleteInbox
      parameters:
        - $ref: "#/components/parameters/inboxId"
      responses:
        "204":
          description: Inbox deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /inboxes/{inbox_id}/messages:
    get:
      tags: [Messages]
      summary: List messages
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/inboxId"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/pageToken"
        - name: thread_id
          in: query
          schema:
            type: string
          description: Filter by thread
        - name: label
          in: query
          schema:
            type: string
          description: Filter by label
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Messages]
      summary: Send message
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/inboxId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /inboxes/{inbox_id}/messages/{message_id}:
    get:
      tags: [Messages]
      summary: Get message
      operationId: getMessage
      parameters:
        - $ref: "#/components/parameters/inboxId"
        - $ref: "#/components/parameters/messageId"
      responses:
        "200":
          description: Message details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Messages]
      summary: Delete message
      operationId: deleteMessage
      parameters:
        - $ref: "#/components/parameters/inboxId"
        - $ref: "#/components/parameters/messageId"
      responses:
        "204":
          description: Message deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "API key in format: oam_xxxxxxxx"
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    pageToken:
      name: page_token
      in: query
      schema:
        type: string
      description: Pagination cursor
    podId:
      name: pod_id
      in: path
      required: true
      schema:
        type: string
      description: Pod identifier
    inboxId:
      name: inbox_id
      in: path
      required: true
      schema:
        type: string
      description: Inbox identifier
    messageId:
      name: message_id
      in: path
      required: true
      schema:
        type: string
      description: Message identifier
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict (duplicate client_id)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - validation_error
                - authentication_error
                - authorization_error
                - not_found
                - conflict
                - rate_limit_exceeded
                - internal_error
            message:
              type: string
            details:
              type: object
...
    Pod:
      type: object
      required: [pod_id, name, created_at, updated_at]
      properties:
        pod_id:
          type: string
        name:
          type: string
        client_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PodList:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Pod"
        next_page_token:
          type: string
          nullable: true
        has_more:
          type: boolean
    CreatePodRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        client_id:
          type: string
          maxLength: 255
    Inbox:
      type: object
      required: [inbox_id, pod_id, username, domain, email, created_at, updated_at]
      properties:
        inbox_id:
          type: string
        pod_id:
          type: string
        username:
          type: string
        domain:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        client_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InboxList:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Inbox"
        next_page_token:
          type: string
          nullable: true
        has_more:
          type: boolean
    CreateInboxRequest:
      type: object
      properties:
        username:
          type: string
          pattern: "^[a-z0-9][a-z0-9._-]*[a-z0-9]$"
          minLength: 2
          maxLength: 64
        domain:
          type: string
        display_name:
          type: string
          maxLength: 255
        client_id:
          type: string
          maxLength: 255
    Attachment:
      type: object
      required: [attachment_id, filename, content_type, size]
      properties:
        attachment_id:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          description: Size in bytes
    AttachmentInput:
      type: object
      required: [filename, content]
      properties:
        filename:
          type: string
        content:
          type: string
          format: byte
          description: Base64-encoded content
        content_type:
          type: string
    Message:
      type: object
      required: [message_id, inbox_id, thread_id, from, to, created_at]
      properties:
        message_id:
          type: string
        inbox_id:
          type: string
        thread_id:
          type: string
        from:
          type: string
        to:
          type: array
          items:
            type: string
        cc:
          type: array
          items:
            type: string
        bcc:
          type: array
          items:
            type: string
        subject:
          type: string
          nullable: true
        text:
          type: string
          nullable: true
        html:
          type: string
          nullable: true
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        labels:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
    MessageList:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        next_page_token:
          type: string
          nullable: true
        has_more:
          type: boolean
    SendMessageRequest:
      type: object
      required: [to, subject]
      properties:
        to:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        cc:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        bcc:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        subject:
          type: string
          maxLength: 998
        text:
          type: string
        html:
          type: string
        reply_to:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentInput"
        labels:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
...